requirements.txt
pandas
numpy
websocket-client
getpass
simplejson
threading